requirements.txt
google-genai
aiosqlite
uvicorn[standard]